+++ service/src/main/kotlin/io/provenance/explorer/grpc/v1/AccountGrpcClient.kt
@@ -1,5 +1,6 @@
 package io.provenance.explorer.grpc.v1
 
+import cosmos.gov.v1beta1.QueryOuterClass
 import io.grpc.ManagedChannelBuilder
 import io.provenance.explorer.config.GrpcLoggingInterceptor
 import io.provenance.explorer.grpc.extensions.getPaginationBuilder
@@ -14,6 +15,8 @@ import cosmos.distribution.v1beta1.QueryGrpc as DistGrpc
 import cosmos.distribution.v1beta1.QueryOuterClass as DistOuterClass
 import cosmos.staking.v1beta1.QueryGrpc as StakingGrpc
 import cosmos.staking.v1beta1.QueryOuterClass as StakingOuterClass
+import cosmos.mint.v1beta1.QueryGrpc as MintGrpc
+import cosmos.mint.v1beta1.QueryOuterClass as MintOuterClass
 
 @Component
 class AccountGrpcClient(channelUri : URI) {
@@ -22,6 +25,7 @@ class AccountGrpcClient(channelUri : URI) {
     private val bankClient: BankQueryGrpc.QueryBlockingStub
     private val stakingClient: StakingGrpc.QueryBlockingStub
     private val distClient: DistGrpc.QueryBlockingStub
+    private val mintClient: MintGrpc.QueryBlockingStub
 
     init {
         val channel =
@@ -43,6 +47,7 @@ class AccountGrpcClient(channelUri : URI) {
         bankClient = BankQueryGrpc.newBlockingStub(channel)
         stakingClient = StakingGrpc.newBlockingStub(channel)
         distClient = DistGrpc.newBlockingStub(channel)
+        mintClient = MintGrpc.newBlockingStub(channel)
     }
 
     fun getAccountInfo(address: String) =
@@ -107,4 +112,9 @@ class AccountGrpcClient(channelUri : URI) {
                 .setDelegatorAddress(delAddr)
                 .build())
 
+    fun getBankParams() = bankClient.params(BankOuterClass.QueryParamsRequest.newBuilder().build())
+
+    fun getAuthParams() = authClient.params(AuthOuterClass.QueryParamsRequest.newBuilder().build())
+
+    fun getMintParams() = mintClient.params(MintOuterClass.QueryParamsRequest.newBuilder().build())
 }