+++ service/src/main/kotlin/io/provenance/explorer/grpc/v1/IbcGrpcClient.kt
@@ -8,6 +8,7 @@ import java.net.URI
 import java.util.concurrent.TimeUnit
 import ibc.applications.transfer.v1.QueryGrpc as TransferQueryGrpc
 import ibc.applications.transfer.v1.QueryOuterClass as TransferOuterClass
+import ibc.core.client.v1.QueryOuterClass as ClientOuterClass
 import ibc.core.channel.v1.QueryGrpc as ChannelQueryGrpc
 import ibc.core.channel.v1.QueryOuterClass as ChannelOuterClass
 import ibc.core.client.v1.QueryGrpc as ClientQueryGrpc
@@ -63,5 +64,7 @@ class IbcGrpcClient(channelUri: URI) {
                 .setChannelId(channel)
                 .build())
 
+    fun getTransferParams() = transferClient.params(TransferOuterClass.QueryParamsRequest.newBuilder().build())
 
+    fun getClientParams() = clientClient.clientParams(ClientOuterClass.QueryClientParamsRequest.newBuilder().build())
 }