+++ service/src/main/kotlin/io/provenance/explorer/grpc/v1/MarkerGrpcClient.kt
@@ -9,6 +9,7 @@ import io.provenance.marker.v1.QueryGrpc
 import io.provenance.marker.v1.QueryHoldingRequest
 import io.provenance.marker.v1.QueryHoldingResponse
 import io.provenance.marker.v1.QueryMarkerRequest
+import io.provenance.marker.v1.QueryParamsRequest as MarkerRequest
 import org.springframework.stereotype.Component
 import java.net.URI
 import java.util.concurrent.TimeUnit
@@ -56,5 +57,5 @@ class MarkerGrpcClient(channelUri: URI) {
             QueryHoldingResponse.getDefaultInstance()
         }
 
-
+    fun getMarkerParams() = markerClient.params(MarkerRequest.newBuilder().build())
 }