+++ service/src/main/kotlin/io/provenance/explorer/service/ExplorerService.kt
@@ -9,14 +9,11 @@ import io.provenance.explorer.domain.entities.TxCacheRecord
 import io.provenance.explorer.domain.extensions.NHASH
 import io.provenance.explorer.domain.extensions.formattedString
 import io.provenance.explorer.domain.extensions.height
+import io.provenance.explorer.domain.extensions.toDecCoin
 import io.provenance.explorer.domain.extensions.toHash
 import io.provenance.explorer.domain.extensions.translateByteArray
-import io.provenance.explorer.domain.models.explorer.BlockSummary
-import io.provenance.explorer.domain.models.explorer.CountStrTotal
-import io.provenance.explorer.domain.models.explorer.CountTotal
-import io.provenance.explorer.domain.models.explorer.DateTruncGranularity
-import io.provenance.explorer.domain.models.explorer.PagedResults
-import io.provenance.explorer.domain.models.explorer.Spotlight
+import io.provenance.explorer.domain.models.explorer.*
+import io.provenance.explorer.grpc.v1.*
 import io.provenance.explorer.service.async.AsyncCaching
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.runBlocking
@@ -25,7 +22,6 @@ import org.springframework.stereotype.Service
 import java.math.BigDecimal
 import java.math.RoundingMode
 
-
 @Service
 class ExplorerService(
     private val props: ExplorerProperties,
@@ -33,7 +29,14 @@ class ExplorerService(
     private val blockService: BlockService,
     private val accountService: AccountService,
     private val validatorService: ValidatorService,
-    private val asyncCaching: AsyncCaching
+    private val asyncCaching: AsyncCaching,
+    private val govClient: GovGrpcClient,
+    private val accountClient: AccountGrpcClient,
+    private val ibcClient: IbcGrpcClient,
+    private val attrClient: AttributeGrpcClient,
+    private val metadataClient: MetadataGrpcClient,
+    private val markerClient: MarkerGrpcClient,
+    private val validatorClient: ValidatorGrpcClient
 ) {
 
     protected val logger = logger(ExplorerService::class)
@@ -116,4 +119,107 @@ class ExplorerService(
         ChainGasFeeCacheRecord.findForDates(fromDate, toDate, count).reversed()
 
     fun getChainId() = asyncCaching.getChainIdString()
+
+    fun getParams() : Params {
+        val authParams = accountClient.getAuthParams().params
+        val bankParams = accountClient.getBankParams().params
+        val distParams = validatorClient.getDistParams().params
+        val votingParams = govClient.getParams(GovParamType.voting).votingParams
+        val tallyParams = govClient.getParams(GovParamType.tallying).tallyParams
+        val depositParams = govClient.getParams(GovParamType.deposit).depositParams
+        val mintParams = accountClient.getMintParams().params
+        val slashingParams = validatorClient.getSlashingParams().params
+        val stakingParams = validatorClient.getStakingParams().params
+        val transferParams = ibcClient.getTransferParams().params
+        val clientParams = ibcClient.getClientParams().params
+        val attrParams = attrClient.getAttrParams().params
+        val markerParams = markerClient.getMarkerParams().params
+        val metadataParams = metadataClient.getMetadataParams().params
+        val nameParams = attrClient.getNameParams().params
+
+        return Params(
+            CosmosParams(
+                AuthParams(
+                    authParams.maxMemoCharacters,
+                    authParams.txSigLimit,
+                    authParams.txSigLimit,
+                    authParams.sigVerifyCostEd25519,
+                    authParams.sigVerifyCostSecp256K1,
+                ),
+                BankParams(
+                    bankParams.defaultSendEnabled,
+                ),
+                DistParams(
+                    distParams.communityTax.toDecCoin(),
+                    distParams.baseProposerReward.toDecCoin(),
+                    distParams.bonusProposerReward.toDecCoin(),
+                    distParams.withdrawAddrEnabled,
+                ),
+                GovParams(
+                    VotingParams(
+                        votingParams.votingPeriod.seconds,
+                    ),
+                    TallyingParams(
+                        tallyParams.quorum.toString(Charsets.UTF_8).toDecCoin(),
+                        tallyParams.threshold.toString(Charsets.UTF_8).toDecCoin(),
+                        tallyParams.vetoThreshold.toString(Charsets.UTF_8).toDecCoin(),
+                    ),
+                    DepositParams(
+                        MinDeposit(
+                            depositParams.getMinDeposit(0).denom,
+                            depositParams.getMinDeposit(0).amount,
+                        ),
+                        depositParams.maxDepositPeriod.seconds,
+                    ),
+                ),
+                MintParams(
+                    mintParams.mintDenom,
+                    mintParams.inflationRateChange.toDecCoin(),
+                    mintParams.inflationMax,
+                    mintParams.inflationMin,
+                    mintParams.goalBonded.toDecCoin(),
+                    mintParams.blocksPerYear,
+                ),
+                SlashingParams(
+                    slashingParams.signedBlocksWindow,
+                    slashingParams.minSignedPerWindow.toString(Charsets.UTF_8).toDecCoin(),
+                    slashingParams.downtimeJailDuration.seconds,
+                    slashingParams.slashFractionDoubleSign.toString(Charsets.UTF_8).toDecCoin(),
+                    slashingParams.slashFractionDowntime.toString(Charsets.UTF_8).toDecCoin(),
+                ),
+                StakingParams(
+                    stakingParams.unbondingTime.seconds,
+                    stakingParams.maxValidators,
+                    stakingParams.maxEntries,
+                    stakingParams.bondDenom,
+                ),
+                IBCParams(
+                    TransferParams(
+                        transferParams.sendEnabled,
+                        transferParams.receiveEnabled,
+                    ),
+                    ClientParams(
+                        clientParams.allowedClientsList,
+                    ),
+                ),
+            ),
+            ProvParams(
+                AttributeParams(
+                    attrParams.maxValueLength,
+                ),
+                MarkerParams(
+                    markerParams.maxTotalSupply,
+                    markerParams.enableGovernance,
+                    markerParams.unrestrictedDenomRegex,
+                ),
+                metadataParams.toString(),
+                NameParams(
+                    nameParams.maxSegmentLength,
+                    nameParams.minSegmentLength,
+                    nameParams.maxNameLevels,
+                    nameParams.allowUnrestrictedNames,
+                ),
+            ),
+        )
+    }
 }