+++ service/src/main/kotlin/io/provenance/explorer/web/v2/GeneralController.kt
@@ -2,6 +2,7 @@ package io.provenance.explorer.web.v2
 
 import io.provenance.explorer.domain.models.explorer.DateTruncGranularity
 import io.provenance.explorer.domain.models.explorer.Spotlight
+import io.provenance.explorer.domain.models.explorer.Params
 import io.provenance.explorer.service.ExplorerService
 import io.swagger.annotations.Api
 import io.swagger.annotations.ApiOperation
@@ -21,6 +22,11 @@ import javax.validation.constraints.Min
 @RequestMapping(path = ["/api/v2"], produces = [MediaType.APPLICATION_JSON_VALUE])
 @Api(value = "General controller", produces = "application/json", consumes = "application/json", tags = ["General"])
 class GeneralController(private val explorerService: ExplorerService) {
+
+    @ApiOperation("Returns parameters for all the modules")
+    @GetMapping("/params")
+    fun param(): ResponseEntity<Params> = ResponseEntity.ok(explorerService.getParams())
+
 
     @ApiOperation("Returns spotlight statistics")
     @GetMapping("/spotlight")